movies-service/etl/requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
elastic-transport==8.4.0
elasticsearch==8.7.0
psycopg2-binary==2.9.5
pydantic==1.10.7
python-dotenv==1.0.0
redis==4.5.4
typing_extensions==4.5.0
urllib3==1.26.15